Revert "Revert "Added a bug-reproducing testcase in MapWalkerTestClass""
This reverts commit 38b4df14a8c912a89ca505aec602f62318df8371.

+++ src/swpg3/game/map/Map.java
@@ -435,7 +435,7 @@ public class Map {
 						}
 					} else if (getTileAt(w, h).getStatus() == TileStatus.EXPANSION && overridePossible)
 					{
-						possibleMoves.add(new Move(pos.clone(), (byte) 0, (byte) playerNumber));
+						possibleMoves.add(new Move(pos.clone(), (byte) 0, playerNumber));
 					}
 				}
 			}
@@ -587,7 +587,7 @@ public class Map {
 						}
 					} else if (getTileAt(w, h).getStatus() == TileStatus.EXPANSION && overridePossible)
 					{
-						possibleMoves.add(new Move(pos.clone(), (byte) 0, (byte) playerNumber, MoveType.OVERRIDE_USE));
+						possibleMoves.add(new Move(pos.clone(), (byte) 0, playerNumber, MoveType.OVERRIDE_USE));
 					}
 				}
 			}
